requirements.txt
# ManualBook - Complete Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies (used across all modules)
requests==2.32.3
python-dotenv==1.0.0

# Translation (Ingress/gtranslate_md.py)
deep-translator==1.11.4

# Document processing (Ingress/)
python-docx==1.1.0

# Vector database (Ingress/vectorize.py)
chromadb==0.5.3

# Backend API server (Backend/)
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.0
